requirements.txt
openai>=1.45.0
pillow>=10.0.0
tornado>=6.3
